requirements.txt
flask~=3.1
requests~=2.32
python-dotenv~=1.0
